removed template builder
